Move Handoff struct back into stage0 to make lib/stage0-handoff platform independent
